Fix non-zero origin in `limit_model_to_receiver_area` func (#113)
* Fixed typo in `limit_model_to_receiver_area` func

Before that `limit_m=true` option sometimes used to lead to exception.
The cause was the typo within `limit_model_to_receiver_area` function.
Now the typo is fixed and model origin is taken into account

* Added dirty test `limit_model_to_receiver_area`

* `limit_model_to_receiver_area` is cleaned

The test was moved from geometry testing file to `test_issues.jl`.
The body of the test is cleaned up

* Style fix

* Limit model test moved to `test_basics.jl`

* Style fix

* Fixed typo